Feat(Plan): plan 달력 api 조회

+++ jabiseo-api/src/main/java/com/jabiseo/plan/dto/calender/PlanProgressResponse.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 @Getter
 public class PlanProgressResponse {
     ActivityType activityType;
-    Integer completedValue;
+    Long completedValue;
     Integer targetValue;
 
-    public PlanProgressResponse(ActivityType activityType, Integer completedValue, Integer targetValue) {
+    public PlanProgressResponse(ActivityType activityType, Long completedValue, Integer targetValue) {
         this.activityType = activityType;
         this.completedValue = completedValue;
         this.targetValue = targetValue;
